Added note regarding .obj file in some rust target builds causing Godot import error
After struggling for a while with my own project, I found out that Godot was attempting to import an object binary file inmy target dir with the extension '.obj' as a wavefront OBJ file, and was failing to build due to a supposed corrupted file. Adding a .gdignore fixed this issue, and I wanted to warn others in case they run into it.