Note regarding .obj files, interpreted as wavefront by Godot (#12)
After struggling for a while with my own project, I found out that Godot was attempting to import an object binary file in my target dir with the extension '.obj' as a wavefront OBJ file, and was failing to build due to a supposed corrupted file. Adding a .gdignore fixed this issue, and I wanted to warn others in case they run into it.