Add missing rustflags for exporting to web
According to [a message in the Discord server](https://discord.com/channels/723850269347283004/1311869218572210207/1312024661860683847), you probably need some [more flags](https://discord.com/channels/723850269347283004/1274080437198651546/1275336695264575542) in order to properly export to the web. In my experience, this makes it work! I have not tested to verify whether this is the exact set needed, or a superset.